separate CloseButton into its own file
